+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 Please file changes under `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed` or `Security`.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
+## Unreleased
+
+## 2019-05-06
+
+### Added
+
+- Added parcelshop locator [@timoj](https://github.com/timoj)
+- Added config for dispatch selection for locator [@timoj](https://github.com/timoj)
+- Added config for en-/disabling the locator [@timoj](https://github.com/timoj)
+- Added checkout validation for the locator [@timoj](https://github.com/timoj)
+
 ## Released
 
 ## [1.2.3](https://github.com/kabisa/wuunder-webshopplugin-shopware/releases/tag/1.2.3) - 2018-11-07
@@ -46,7 +57,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Fix
 
 - Fixed folder structure plugin [@timoj](https://github.com/timoj)
-
 
 # Old change log